+++ CHANGELOG.md
@@ -5,7 +5,16 @@ We are currently working on porting this changelog to the specifications in
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## Version 0.1.8 - Unreleased
+## Version 0.1.9 - Unreleased
+
+### Added
+* Support for more sbatch options in slurm backend
+
+### Fixed
+* Bugs in slurm backend
+
+
+## Version 0.1.8 - Released 2023-03-05
 
 ### Added:
 * New experimental CLI-queue feature. Create a pipeline in bash using the CLI.